finalizada o cadastro modelo pergunta

+++ database/migrations/2024_10_23_030139_create_medidas_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('medidas', function (Blueprint $table) {
             $table->id();
-            $table->integer('peso');
+            $table->decimal('peso', 10, 2);
             $table->date('data');
             $table->unsignedBigInteger('cliente_id')->unsigned();
             $table->foreign('cliente_id')->references('id')->on('clientes')->nullable();
